# openapi.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "SolTerm API",
      "version": "1.0.0",
      "description": "Comprehensive API for SolTerm SSH management, file operations, tunneling, and server monitoring. This API provides endpoints for managing SSH hosts, file operations, tunnel connections, server monitoring, user management, and system alerts.",
      "contact": {
        "name": "SolTerm Development Team"
      }
    },
    "servers": [
      {
        "url": "http://localhost:8081",
        "description": "Main database and authentication server"
      },
      {
        "url": "http://localhost:8083",
        "description": "SSH tunnel management server"
      },
      {
        "url": "http://localhost:8084",
        "description": "SSH file manager server"
      },
      {
        "url": "http://localhost:8085",
        "description": "Server statistics and monitoring server"
      }
    ],
    "security": [
      {
        "bearerAuth": []
      }
    ],
...
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      },
      "schemas": {
...
        "ServerStatus": {
          "type": "object",
          "properties": {
            "status": { "type": "string", "enum": ["online", "offline"] },
            "lastChecked": { "type": "string", "format": "date-time" }
          }
        },
...
      },
      "parameters": {
...
        "sessionId": {
          "name": "sessionId",
          "in": "query",
          "description": "The SSH session identifier",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
...
        "serverIdPath": {
          "name": "id",
          "in": "path",
          "description": "The server identifier",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      },
...
    },
    "paths": {
...
             "/ssh/file_manager/ssh/status": {
         "get": {
           "summary": "Get SSH connection status",
           "description": "Get SSH connection status. **Server: localhost:8084**",
          "operationId": "getSSHStatus",
          "tags": ["SSH File Operations"],
          "parameters": [
            {
              "$ref": "#/components/parameters/sessionId"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "connected": { "type": "boolean" }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
             "/status/{id}": {
         "get": {
           "summary": "Get server status by ID",
           "description": "Get server status by ID. **Server: localhost:8085**",
          "operationId": "getServerStatusById",
          "tags": ["Server Statistics"],
          "parameters": [
            {
              "$ref": "#/components/parameters/serverIdPath"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ServerStatus" }
                }
              }
            }
          }
        }
      },
...
    }
  }